rename : pages 폴더 생성, 상품 페이지 pages 폴더로 이동
