+++ src/components/App.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import Navbar from "./Navbar";
 import { Route, Routes, BrowserRouter } from "react-router-dom";
-import ItemPage from "./ItemPage";
+import ItemPage from "../pages/ItemPage";
 import Home from "./Home";
 
 function App() {
